requirements.txt
anthropic>=0.40.0
praw>=7.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
geopy>=2.4.0
rich>=13.7.0
click>=8.1.0
